added firebase interactor
