+++ firebase/firebaseInteractor.ts
@@ -1,0 +1,20 @@
+import * as firebase from "firebase/app";
+import "firebase/auth";
+
+
+// Your web app's Firebase configuration
+const firebaseConfig = {
+    apiKey: "AIzaSyC9cAIskIjacQRxhbJlJxJNizGOlbRj4gk",
+    authDomain: "flowinthefield.firebaseapp.com",
+    projectId: "flowinthefield",
+    storageBucket: "flowinthefield.appspot.com",
+    messagingSenderId: "788116866647",
+    appId: "1:788116866647:web:0e2cbfc671576c8089c512"
+};
+
+// Initialize Firebase
+const app = firebase.initializeApp(firebaseConfig);
+
+export default class FirebaseInteractor {
+    auth = firebase.auth();
+}
